test: 서비스단 단위 테스트 추가
