+++ backend/src/main/java/com/wootech/dropthecode/controller/AuthController.java
@@ -12,7 +12,10 @@ import com.wootech.dropthecode.service.OauthService;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class AuthController {
